+++ src/main/kotlin/de/craftan/engine/actions/RollDiceAction.kt
@@ -3,6 +3,8 @@ package de.craftan.engine.actions
 import de.craftan.CraftanGameAction
 import de.craftan.engine.*
 import de.craftan.engine.events.actions.RolledDiceEvent
+import de.craftan.engine.gameflow.CraftanActionItem
+import de.craftan.engine.gameflow.craftanActionItem
 import de.craftan.engine.map.DiceNumber
 import org.bukkit.Material
 import kotlin.random.Random
