+++ src/main/kotlin/de/craftan/engine/gameflow/states/DecidableState.kt
@@ -1,4 +1,7 @@
-package de.craftan.engine
+package de.craftan.engine.gameflow.states
+
+import de.craftan.engine.CraftanGame
+import de.craftan.engine.gameflow.RoundState
 
 /**
  * Can be used to manipulate the flow of a round
