+++ src/main/kotlin/de/craftan/engine/map/graph/Node.kt
@@ -1,8 +1,8 @@
 package de.craftan.engine.map.graph
 
-import de.craftan.engine.CraftanStructure
 import de.craftan.engine.map.TileCoordinate
 import de.craftan.engine.map.TileDirection
+import de.craftan.engine.structures.CraftanStructure
 
 /**
  * A Node is practically the place where the players can build settlements and cities.
