+++ src/main/kotlin/de/craftan/engine/structures/Settlement.kt
@@ -1,7 +1,6 @@
 package de.craftan.engine.structures
 
 import de.craftan.engine.CraftanResource
-import de.craftan.engine.CraftanStructure
 import de.craftan.engine.resources.WoodResource
 
 class Settlement : CraftanStructure {
